Rename and move basicmacros.h to util/basic_macros.h
