+++ src/threading/mutex.h
@@ -44,7 +44,7 @@ DEALINGS IN THE SOFTWARE.
 	#include <pthread.h>
 #endif
 
-#include "basicmacros.h"
+#include "util/basic_macros.h"
 
 class Mutex
 {
